fixed a lot of stuff
including video playback, back from activity, etc..
