+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,8 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme">
+        android:theme="@style/AppTheme"
+        android:name=".MainActivity">
         <activity
             android:name=".Login.LoginSignupVideoBgPage"
             android:label="@string/app_name"
@@ -23,8 +24,8 @@
         </activity>
 
         <activity android:name=".Login.Login"/>
-        <activity android:name=".MainActivity"/>
         <activity android:name=".Login.Signup"/>
+        <activity android:name=".BidBoard"/>
 
     </application>
