+++ app/src/main/java/com/gimmyo/car/app/BidBoard.java
@@ -1,0 +1,40 @@
+package com.gimmyo.car.app;
+
+import android.os.Bundle;
+import android.support.design.widget.FloatingActionButton;
+import android.support.design.widget.Snackbar;
+import android.support.v7.app.AppCompatActivity;
+import android.view.View;
+import android.webkit.WebView;
+import android.webkit.WebViewClient;
+
+/**
+ * Created by jonegalado on 7/27/17, Gimmyo Project.
+ */
+
+public class BidBoard extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        WebView myWebView = (WebView) findViewById(R.id.webview);
+        myWebView.loadUrl("http://www.truecar.com");
+        myWebView.setWebViewClient(new WebViewClient());
+
+        FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
+        fab.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
+                        .setAction("Action", null).show();
+            }
+        });
+    }
+
+    public void onBackPressed(){
+        super.onBackPressed();
+        this.finish();
+    }
+}
