+++ app/src/main/java/com/gimmyo/car/app/Login/LoginSignupVideoBgPage.java
@@ -3,7 +3,9 @@ package com.gimmyo.car.app.Login;
 import android.content.Intent;
 import android.media.MediaPlayer;
 import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
+import android.support.annotation.RequiresApi;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
@@ -24,7 +26,17 @@ public class LoginSignupVideoBgPage extends AppCompatActivity implements View.On
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.video_bg_page_login_signup);
+        runVideo();
 
+        Button loginButton = (Button) findViewById(R.id.gimmyLoginButton);
+        loginButton.setOnClickListener(this);
+
+        Button signupButton = (Button) findViewById(R.id.signupGimmyoButton);
+        signupButton.setOnClickListener(this);
+
+    }
+
+    public void runVideo (){
         mVideoView = (VideoView) findViewById(R.id.bgVideoView);
 
         Uri uri = Uri.parse("android.resource://"+getPackageName()+"/"+R.raw.bg_video);
@@ -39,13 +51,6 @@ public class LoginSignupVideoBgPage extends AppCompatActivity implements View.On
                 mediaPlayer.setLooping(true);
             }
         });
-
-        Button loginButton = (Button) findViewById(R.id.gimmyLoginButton);
-        loginButton.setOnClickListener(this);
-
-        Button signupButton = (Button) findViewById(R.id.signupGimmyoButton);
-        signupButton.setOnClickListener(this);
-
     }
 
     @Override
@@ -62,5 +67,18 @@ public class LoginSignupVideoBgPage extends AppCompatActivity implements View.On
                 startActivity(signupIntent);
                 break;
         }
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        runVideo();
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.JELLY_BEAN)
+    @Override
+    public void onBackPressed() {
+        super.onBackPressed();
+        this.finishAffinity();
     }
 }
